+++ src/main/java/service/CartService.java
@@ -101,7 +101,7 @@ public class CartService implements ICart {
                     listItem.add(new CartItem(new Product(
                             resultSet.getInt("id"),
                             resultSet.getString("type"),
-                            resultSet.getFloat("value"),
+                            resultSet.getInt("value"),
                             resultSet.getInt("rate"),
                             resultSet.getString("currency"),
                     resultSet.getString("image")
